Add type check to gate, split type check
