+++ .github/workflows/main.yml
@@ -24,12 +24,13 @@ jobs:
     strategy:
       matrix:
         task:
-          - check-whole-repo-tests
-          - lint
-          - check-format
           - tox-cover
           - tox-nocover
           - tox-rest
+          - check-whole-repo-tests
+          - check-types-basic
+          - lint
+          - check-format
       fail-fast: false
     env:
       PYTHON_VERSION: "3.10"
@@ -65,7 +66,8 @@ jobs:
       matrix:
         task:
           - check-documentation
-          - check-types
+          - check-types-mypy
+          - check-types-pyright
           - check-coverage
           - check-conjecture-coverage
           - check-py39-cover
